+++ libraries/DAC8574/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2024-12-30
+- fix #4, add support for A2 and A3 extended address pins.
+  - **setExtendedAddress()** + **getExtendedAddress()**
+  - improve DAC8574_performance.ino
+- correct clearing internal **control** bits, in case power down.
+- update keywords
+- update unit tests
+- update readme.md
+
+
 ## [0.1.0] - 2024-11-11
 - initial version (based upon DAC8571)